Adds support for relations
